flatten package structure
