+++ zio-pubsub-google-test/src/test/scala/com/anymindgroup/pubsub/google/AvroPublisherSpec.scala
@@ -1,12 +1,9 @@
 package com.anymindgroup.pubsub.google
 
+import com.anymindgroup.pubsub.*
 import com.anymindgroup.pubsub.PubsubTestSupport.*
-import com.anymindgroup.pubsub.google
 import com.anymindgroup.pubsub.google.TestSupport.*
-import com.anymindgroup.pubsub.model.*
-import com.anymindgroup.pubsub.pub.*
 import com.anymindgroup.pubsub.serde.VulcanSerde
-import com.anymindgroup.pubsub.sub.*
 import com.google.cloud.pubsub.v1.SubscriptionAdminClient
 import com.google.pubsub.v1.{SubscriptionName, TopicName}
