+++ zio-pubsub-google/src/main/scala/com/anymindgroup/pubsub/google/PublisherConfig.scala
@@ -1,6 +1,6 @@
 package com.anymindgroup.pubsub.google
 
-import com.anymindgroup.pubsub.model.*
+import com.anymindgroup.pubsub.{Encoding, TopicName}
 import com.google.pubsub.v1.TopicName as GTopicName
 
 final case class PublisherConfig(
