+++ zio-pubsub-google/src/main/scala/com/anymindgroup/pubsub/google/package.scala
@@ -1,6 +1,5 @@
 package com.anymindgroup.pubsub
 
-import com.anymindgroup.pubsub.sub.AckReply
 import com.google.pubsub.v1.ReceivedMessage as GReceivedMessage
 
 import zio.stream.ZStream
