+++ zio-pubsub-serde-circe/shared/src/main/scala/com/anymindgroup/pubsub/serde/CirceSerde.scala
@@ -1,11 +1,12 @@
 package com.anymindgroup.pubsub.serde
 
-import com.anymindgroup.pubsub.sub.ReceivedMessage
+import com.anymindgroup.pubsub.{ReceivedMessage, Serde}
 import io.circe.Codec
 import io.circe.parser.decode
 
 import zio.{RIO, ZIO}
 
+@deprecated("will be removed in future release in favor of zio-schema module")
 object CirceSerde {
   def fromCirceCodec[T](codec: Codec[T]): Serde[Any, T] = new Serde[Any, T] {
     private implicit val c: Codec[T] = codec
