+++ zio-pubsub-serde-vulcan/src/main/scala/com/anymindgroup/pubsub/serde/VulcanSerde.scala
@@ -1,11 +1,11 @@
 package com.anymindgroup.pubsub.serde
 
-import com.anymindgroup.pubsub.model.Encoding
-import com.anymindgroup.pubsub.sub.ReceivedMessage
+import com.anymindgroup.pubsub.{Encoding, ReceivedMessage, Serde}
 import vulcan.Codec
 
 import zio.{RIO, ZIO}
 
+@deprecated("will be removed in future release in favor of zio-schema module")
 object VulcanSerde {
   def fromAvroCodec[T](codec: Codec[T], encoding: Encoding): Serde[Any, T] = new Serde[Any, T] {
     private implicit val c: Codec[T] = codec
