+++ zio-pubsub-serde-zio-schema/shared/src/main/scala/com/anymindgroup/pubsub/serde/BinaryCodecSerde.scala
@@ -1,6 +1,4 @@
-package com.anymindgroup.pubsub.serde
-
-import com.anymindgroup.pubsub.sub.ReceivedMessage
+package com.anymindgroup.pubsub
 
 import zio.schema.codec.BinaryCodec
 import zio.{Chunk, RIO, ZIO}
