+++ zio-pubsub-test/shared/src/test/scala/com/anymindgroup/pubsub/SerdesSpec.scala
@@ -1,6 +1,4 @@
-package com.anymindgroup.pubsub.serde
-
-import com.anymindgroup.pubsub.sub.ReceivedMessage
+package com.anymindgroup.pubsub
 
 import zio.Random
 import zio.test.*
