+++ zio-pubsub-test/shared/src/test/scala/com/anymindgroup/pubsub/http/PubAndSubSpec.scala
@@ -4,10 +4,8 @@ import scala.util.Random
 
 import com.anymindgroup.gcp.auth.TokenProvider
 import com.anymindgroup.http.httpBackendScoped
+import com.anymindgroup.pubsub.*
 import com.anymindgroup.pubsub.PubsubTestSupport.*
-import com.anymindgroup.pubsub.model.{SubscriptionName, TopicName}
-import com.anymindgroup.pubsub.pub.PublishMessage
-import com.anymindgroup.pubsub.{Publisher, PubsubConnectionConfig, Serde}
 import sttp.client4.Backend
 
 import zio.test.*
