+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub.scala
@@ -1,29 +1,4 @@
 package com.anymindgroup.pubsub
 
-export model.MessageId
-export model.Topic
-export model.Encoding
-export model.OrderingKey
-export model.SchemaSettings
-export model.SchemaRegistry
-export model.PubsubConnectionConfig
-export model.GcpProject
-export model.TopicName
-export model.SubscriptionName
-
-export pub.Publisher
-export pub.PublishMessage
-
-export sub.Subscriber
-export sub.Subscription
-
-export sub.SubscriberFilter
-export sub.ReceivedMessage
-export sub.AckReply
-export sub.AckId
-export sub.DeadLettersSettings
-
-export serde.Deserializer
-export serde.Serializer
-export serde.Serde
-export serde.Serdes
+type Receipt[E] = (ReceivedMessage[E], AckReply)
+type RawReceipt = Receipt[Array[Byte]]
