+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/MessageId.scala
@@ -1,4 +1,4 @@
-package com.anymindgroup.pubsub.model
+package com.anymindgroup.pubsub
 
 opaque type MessageId = String
 object MessageId:
