+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/OrderingKey.scala
@@ -1,7 +1,6 @@
-package com.anymindgroup.pubsub.model
+package com.anymindgroup.pubsub
 
 opaque type OrderingKey = String
-
 object OrderingKey:
   def fromString(o: String): Option[OrderingKey] = Option(o).filter(_.nonEmpty)
