+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/PubsubConnectionConfig.scala
@@ -1,0 +1,5 @@
+package com.anymindgroup.pubsub
+
+enum PubsubConnectionConfig:
+  case Cloud
+  case Emulator(host: String, port: Int = 8085)
