+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/Serde.scala
@@ -1,6 +1,4 @@
-package com.anymindgroup.pubsub.serde
-
-import com.anymindgroup.pubsub.sub.ReceivedMessage
+package com.anymindgroup.pubsub
 
 import zio.RIO
