+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/SubscriptionName.scala
@@ -1,4 +1,4 @@
-package com.anymindgroup.pubsub.model
+package com.anymindgroup.pubsub
 
 final case class SubscriptionName(projectId: String, subscription: String) {
   def fullName: String          = s"projects/$projectId/subscriptions/$subscription"
