+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/Topic.scala
@@ -1,5 +1,3 @@
-package com.anymindgroup.pubsub.model
-
-import com.anymindgroup.pubsub.serde.Serde
+package com.anymindgroup.pubsub
 
 final case class Topic[R, T](name: TopicName, schemaSetting: SchemaSettings, serde: Serde[R, T])
