+++ zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/TopicName.scala
@@ -1,4 +1,4 @@
-package com.anymindgroup.pubsub.model
+package com.anymindgroup.pubsub
 
 final case class TopicName(projectId: String, topic: String) {
   def fullName: String          = s"projects/$projectId/topics/$topic"
