--- zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/sub.scala
@@ -1,4 +1,0 @@
-package com.anymindgroup.pubsub.sub
-
-type Receipt[E] = (ReceivedMessage[E], AckReply)
-type RawReceipt = Receipt[Array[Byte]]
