--- zio-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/sub/AckReply.scala
@@ -1,10 +1,0 @@
-package com.anymindgroup.pubsub.sub
-
-import zio.UIO
-
-final case class AckId(value: String) extends AnyVal
-
-trait AckReply {
-  def ack(): UIO[Unit]
-  def nack(): UIO[Unit]
-}
